requirements.txt
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
pandas>=1.3.0
matplotlib>=3.3.0
Pillow>=8.3.0
tqdm>=4.60.0
PyYAML>=5.4.0
transformers>=4.20.0